+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next release
 
+### Added
+
+- Add new topic/tag filter on homepage.
+
 ## 2022-03-22
 
 ### Fixed